wizAPI.py update
I have fixed the face_arrow() method in the wizAPI, the reason being as the bot would not be able to exit the dungeon if the player was not in the first slot, hence if a random player joins and takes that spot, the bot is forever stuck as the rotation was based off timing for the first slot.

face_arrow() was fixed to bypass this issue, it should work to exit dungeon for the 1st and 2nd slot, but I have not tested it for the 3rd and forth.