replaced feature
wizAPI came with a method that uses your potion when you're low on mana, it was not functional, so i made my own.

the bot now checks if you have low mana when the queue for the dungeon starts. if you have low mana it will use a potion once and wait out the remaining time.

await_finished_loading() now has a paremeter which is a time buffer. reason being if the player is not in the loading screen by the time the bot calls it, the method is skipped as the two while statements aren't called and the rest of the code continues when it shouldn't as it is meant to run after the loading.